requirements.txt
pymupdf
elasticsearch
sentence_transformers
openai
python-dotenv
